Pass 21a: Remove IDE0004/IDE0005 csproj suppressions
Drop `IDE0004` (cast is redundant) and `IDE0005` (using directive is unnecessary)
from the writer's `<NoWarn>` list and fix the 5 violations they exposed:

- `Helpers/CodeWriters.Guids.cs` switch-expression default arms had explicit
  `(ushort)0` and `(byte)0` casts that the analyzer flagged as redundant.
- `Extensions/TypeSignatureExtensions.cs` had an unused
  `using AsmResolver.DotNet;` directive (the file only uses
  `AsmResolver.DotNet.Signatures` and `AsmResolver.PE.DotNet.Metadata.Tables`).
- `Factories/CodeWriters.MappedInterfaceStubs.cs` had an unused
  `using AsmResolver.DotNet;` directive.
- `Models/MethodSignatureInfo.cs` had an unused
  `using AsmResolver.DotNet.Collections;` directive.

Validation: builds clean (0 warnings, 0 errors), all 8 regen scenarios
produce byte-identical output to baseline.

+++ src/WinRT.Projection.Writer/Helpers/CodeWriters.Guids.cs
@@ -80,14 +80,14 @@ internal static partial class CodeWriters
             ushort u => u,
             short s => (ushort)s,
             int i => (ushort)i,
-            _ => (ushort)0
+            _ => 0
         };
         static byte ToByte(object? v) => v switch
         {
             byte b => b,
             sbyte sb => (byte)sb,
             int i => (byte)i,
-            _ => (byte)0
+            _ => 0
         };
     }
     public static void WriteGuid(TextWriter w, TypeDefinition type, bool lowerCase)
